% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_attachments.R
\name{extract_attachments}
\alias{extract_attachments}
\title{Extract attachments embedded in a PDF document}
\usage{
extract_attachments(pdf_path, output_path)
}
\arguments{
\item{pdf_path, output_path}{paths (they get processed by [path.expand()]) for
the input source PDF document and the directory where they should be
extracted to.}
}
\description{
Extract attachments embedded in a PDF document
}
\examples{
pdf_ex <- system.file("extdata", "LeedsBinsOctober2017_Summary.pdf", package="pdfbox")
td <- tempdir()
extract_attachments(pdf_ex, td)
}
