% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-text.R
\name{extract_text}
\alias{extract_text}
\title{Extract text from a PDF document}
\usage{
extract_text(path)
}
\arguments{
\item{path}{path to PDF file (is auto-expanded with \code{\link[=path.expand]{path.expand()}})}
}
\value{
data frame
}
\description{
Extract text from a PDF document
}
\examples{
extract_text(
  system.file(
    "extdata", "imperfect-forward-secrecy-ccs15.pdf", package="pdfbox"
  )
)
}
