% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-uris.r
\name{extract_uris}
\alias{extract_uris}
\title{Extract URI annotations from a PDF document}
\usage{
extract_uris(path)
}
\arguments{
\item{path}{path to PDF file (is auto-expanded with \code{\link[=path.expand]{path.expand()}})}
}
\value{
\code{NULL} or a \code{data.frame}
}
\description{
Extract URI annotations from a PDF document
}
\examples{
extract_uris(
  system.file(
    "extdata",
    "imperfect-forward-secrecy-ccs15.pdf",
    package="pdfbox"
  )
)
}
