% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-images.R
\name{image_count}
\alias{image_count}
\title{Count number of images in a PDF document}
\usage{
image_count(path)
}
\arguments{
\item{path}{path to PDF file (is auto-expanded with \code{\link[=path.expand]{path.expand()}})}
}
\value{
numeric
}
\description{
Count number of images in a PDF document
}
