% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{pdf_info}
\alias{pdf_info}
\title{Retrieve PDF Metadata}
\usage{
pdf_info(path)
}
\arguments{
\item{path}{path to PDF file (is auto-expanded with \code{\link[=path.expand]{path.expand()}})}
}
\value{
data.frame
}
\description{
TODO metadata
}
\examples{
pdf_info(
  system.file(
    "extdata", "imperfect-forward-secrecy-ccs15.pdf", package="pdfbox"
  )
)
}
