% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfbox-package.R
\docType{package}
\name{pdfbox}
\alias{pdfbox}
\alias{pdfbox-package}
\title{Create, Maniuplate and Extract Data from PDF Files}
\description{
Create, Maniuplate and Extract Data from PDF Files
}
\author{
Bob Rudis (bob@rud.is)
}
