% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{stop_logging}
\alias{stop_logging}
\title{Stop logging}
\usage{
stop_logging()
}
\value{
\code{NULL}, invisibly.
}
\description{
Toggle verbose rJava logging
}
\details{
This function turns off the somewhat verbose rJava logging, most of which is
uninformative. It is called automatically when tabulizer is attached via
\code{library()}, \code{require}, etc. To keep logging on, load the package
namespace using \code{requireNamespace("pdfbox")} and reference functions
in using fully qualified references.
}
\note{
This resets a global Java setting and may affect logging of other rJava operations,
requiring a restart of R.
}
\examples{
\dontrun{
stop_logging()
}
}
\author{
Thomas J. Leeper \href{mailto:thosjleeper@gmail.com}{thosjleeper@gmail.com}
}
\keyword{internal}
